# backend/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Auth module REST API. GraphQL is the primary API per the project constitution; this OpenAPI spec covers the REST surface (auth, integrations, webhooks).",
    "title": "Solobueno ERP API",
    "contact": {},
    "version": "0.1"
  },
  "basePath": "/api/v1",
  "paths": {
    "/auth/change-password": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Change the authenticated user's password. Invalidates all other sessions (FR-014).",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "Change password",
        "parameters": [
          {
            "description": "Current and new password",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.ChangePasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.MessageResponse"
            }
          },
          "400": {
            "description": "current_password_incorrect, password_weak",
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.ErrorResponse"
            }
          },
          "401": {
            "description": "unauthorized",
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.ErrorResponse"
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "description": "Authenticate with email and password. If the user belongs to multiple tenants and none is specified, returns 400 tenant_required with the list of choices.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "Log in",
        "parameters": [
          {
            "description": "Login credentials",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.LoginResponse"
            }
          },
          "400": {
            "description": "tenant_required",
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.TenantRequiredResponse"
            }
          },
          "401": {
            "description": "invalid_credentials, account_disabled",
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.ErrorResponse"
            }
          },
          "423": {
            "description": "account_locked",
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.ErrorResponse"
            }
          },
          "429": {
            "description": "rate_limit_exceeded",
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.ErrorResponse"
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Revoke the current session's refresh token. Idempotent - always returns 204.",
        "consumes": ["application/json"],
        "tags": ["auth"],
        "summary": "Log out",
        "parameters": [
          {
            "description": "Refresh token to revoke",
            "name": "request",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.RefreshRequest"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Return the authenticated user's identity and current tenant/role, from the access token claims.",
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "Get current user",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.MeResponse"
            }
          },
          "401": {
            "description": "unauthorized",
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.ErrorResponse"
            }
          }
        }
      }
    },
    "/auth/password-reset/complete": {
      "post": {
        "description": "Set a new password using a valid, unexpired, unused reset token (1-hour TTL). Invalidates all sessions.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "Complete password reset",
        "parameters": [
          {
            "description": "Reset token and new password",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.PasswordResetCompleteRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.MessageResponse"
            }
          },
          "400": {
            "description": "token_invalid, token_expired, token_used, password_weak",
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.ErrorResponse"
            }
          }
        }
      }
    },
    "/auth/password-reset/request": {
      "post": {
        "description": "Send a password reset token to the given email if it exists. Always returns 202 to prevent email enumeration.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "Request password reset",
        "parameters": [
          {
            "description": "Email",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.PasswordResetRequest"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.MessageResponse"
            }
          },
          "429": {
            "description": "rate_limit_exceeded",
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.ErrorResponse"
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Exchange a valid refresh token for a new access/refresh token pair (rotates the refresh token).",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "Refresh access token",
        "parameters": [
          {
            "description": "Refresh token",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.RefreshRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.TokenResponse"
            }
          },
          "401": {
            "description": "token_invalid, session_revoked, token_expired, account_disabled",
            "schema": {
              "$ref": "#/definitions/internal_auth_handler.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "internal_auth_handler.ChangePasswordRequest": {
      "type": "object",
      "properties": {
        "current_password": {
          "type": "string"
        },
        "new_password": {
          "type": "string"
        }
      }
    },
...
    "internal_auth_handler.ErrorDetail": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "locked_until": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "retry_after": {
          "type": "integer"
        },
        "tenants": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/internal_auth_handler.TenantOption"
          }
        }
      }
    },
    "internal_auth_handler.ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/internal_auth_handler.ErrorDetail"
        }
      }
    },
    "internal_auth_handler.LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "tenant_id": {
          "type": "string"
        }
      }
    },
    "internal_auth_handler.LoginResponse": {
      "type": "object",
      "properties": {
        "access_token": {
          "type": "string"
        },
        "expires_at": {
          "type": "string"
        },
        "expires_in": {
          "type": "integer"
        },
        "refresh_token": {
          "type": "string"
        },
        "token_type": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/internal_auth_handler.UserResponse"
        }
      }
    },
    "internal_auth_handler.MeResponse": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "first_name": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "must_reset_password": {
          "type": "boolean"
        },
        "role": {
          "type": "string"
        },
        "tenant_id": {
          "type": "string"
        },
        "tenant_name": {
          "type": "string"
        },
        "tenants": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/internal_auth_handler.TenantRoleInfo"
          }
        }
      }
    },
    "internal_auth_handler.MessageResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "internal_auth_handler.PasswordResetCompleteRequest": {
      "type": "object",
      "properties": {
        "new_password": {
          "type": "string"
        },
        "token": {
          "type": "string"
        }
      }
    },
    "internal_auth_handler.PasswordResetRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "internal_auth_handler.RefreshRequest": {
      "type": "object",
      "properties": {
        "refresh_token": {
          "type": "string"
        }
      }
    },
    "internal_auth_handler.TenantOption": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "slug": {
          "type": "string"
        }
      }
    },
    "internal_auth_handler.TenantRequiredResponse": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/internal_auth_handler.ErrorDetail"
        }
      }
    },
    "internal_auth_handler.TenantRoleInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      }
    },
    "internal_auth_handler.TokenResponse": {
      "type": "object",
      "properties": {
        "access_token": {
          "type": "string"
        },
        "expires_at": {
          "type": "string"
        },
        "expires_in": {
          "type": "integer"
        },
        "refresh_token": {
          "type": "string"
        },
        "token_type": {
          "type": "string"
        }
      }
    },
...
    "internal_auth_handler.UserResponse": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "first_name": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "is_active": {
          "type": "boolean"
        },
        "last_name": {
          "type": "string"
        },
        "must_reset_password": {
          "type": "boolean"
        },
        "role": {
          "type": "string"
        },
        "tenant_id": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Type \"Bearer\" followed by a space and the access token.",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}